fix: show datalayer popover with css/hover instead of Bootrap popover
The Bootstrap Popover API didn't allow to show the Popup on hover/click *and* to also tio select/click links in the popover while also dissming the the popup with leaving/clicking somehwere else on the page.

So for now it's just a simple popover with css hover/tap  state visibillity. The Browser Popover API should be evaluated further, but it was not immediatly clear how to accomplish the above described scenary directly with this api