fix: capitalize language names in dropdown
The local_name for Djangos language names is int the case the respective language would use for language names inside a sentence.

I.e. espanol is lowercase. For showing the langugae names in the dropdown we capitalize the first letter, so it's consistent.

See also Django ticket regarding this: https://code.djangoproject.com/ticket/15714