fix: cast admin column to int/null
In case not all shapes in an import have an integer admin level, pandas would cast them to float which would fail in the to_postgis import due to the schme beeing smallint. Explicit casting in the DataFrame makes sure that the values are either an int or NULL during import to the database.